requirements.txt
mysql-connector-python
SQLAlchemy
strawberry-graphql[fastapi]
fastapi
